fundamental:record: use int64_t for timestamp_delta
Kafka DefaultRecord.java and, librdkafka and other implementations
rely on the value having enough bits to store a long (int64_t)

public class DefaultRecord implements Record {

    private final int sizeInBytes;
    private final byte attributes;
    private final long offset;
    private final long timestamp;
    private final int sequence;
    private final ByteBuffer key;
    private final ByteBuffer value;
    private final Header[] headers;

}

We leave `long offset` as int32_t for now since this is unrelated
to the crash of test=0054 in librdkafka

see [ch934]

+++ src/v/storage/parser_utils.cc
@@ -48,7 +48,7 @@ model::record do_parse_one_record_from_buffer(
     return model::record(
       record_size,
       model::record_attributes(attr),
-      static_cast<int32_t>(timestamp_delta),
+      static_cast<int64_t>(timestamp_delta),
       static_cast<int32_t>(offset_delta),
       key_length,
       std::move(key),
